medworkers: move hospital-data device code to its own TU (fix matrix CUDA Release crash)
The medical-workers init/reroute device lambdas shared a translation unit with generateCellData. On the matrix build (CUDA + RDC + -maxrregcount=255 + fast-math + -O3) the added device code perturbed the Release register allocation of generateCellData's kernel, faulting in writePlotFile -- only on matrix, only in Release (Debug was clean), and only after these gated-off additions (35cdfd3 was clean). Move readHospital* + initHospitalCapacityModel + the reroute (now rerouteHospitalizedToHospital) into AgentContainerHospital.cpp so generateCellData's compilation is no longer affected. No functional change; builds clean.
